Only collection owned bt museum are shown in Showcase selection.
